python_client/pyproject.toml
[project]
name = "python-client"
version = "0.1.0"
description = ""
authors = [
    {name = "Noan Cloarec",email = "noan.cloarec-ext@rca.fr"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "eyed3 (>=0.9.8,<0.10.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "py-picotts (>=0.1.2,<0.2.0)",
]

[project.scripts]
upload_podcasts = 'python_client.upload_podcasts:upload_podcasts'
split_podcasts = 'python_client.split_podcasts:split_podcasts'

[tool.poetry]
packages = [{include = "python_client", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.9.0"
ruff = "^0.14.1"
mypy = "^1.18.2"
pytest-mock = "^3.15.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
